move code from __init__.py to main.py to make import quicker and make version work without dependencies
