+++ .gitignore
@@ -2,7 +2,7 @@ fonts
 *.pdf
 *.pyc
 generated
-dominiontabs.egg-info
+*.egg-info
 build
 dist
 *~
